@@ -7,7 +7,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -20,16 +20,16 @@ dependencies:
   provider: ^6.0.2
   equatable: ^2.0.3
   connectivity_plus: ^2.3.0
-  shared_preferences: ^2.0.13
+  shared_preferences: ^2.0.15
 
   bloc: ^8.0.3
   flutter_bloc: ^8.0.1
   bloc_concurrency: ^0.2.0
 
-  firebase_core: ^1.15.0
-  firebase_analytics: ^9.1.6
-  firebase_crashlytics: ^2.6.3
-  firebase_remote_config: ^2.0.5
+  firebase_core: ^1.16.0
+  firebase_analytics: ^9.1.7
+  firebase_crashlytics: ^2.7.2
+  firebase_remote_config: ^2.0.6
 
 dev_dependencies:
   flutter_test:
@@ -37,8 +37,7 @@ dev_dependencies:
 
   mocktail: ^0.3.0
   bloc_test: ^9.0.3
-  flutter_lints: ^1.0.0
-
+  flutter_lints: ^2.0.0
 
 flutter:
   uses-material-design: true
