@@ -11,23 +11,22 @@ class Navigation {
       NavigationSingleton.instance.navigator;
 
   static Route<dynamic> onGenerateRoute(RouteSettings settings) {
-    // ignore: prefer_typing_uninitialized_variables
-    late final Widget _widget;
+    late final Widget widget;
     // ignore: unused_local_variable
-    late final _args = settings.arguments;
+    late final args = settings.arguments;
 
     switch (settings.name) {
       case SplashScreen.pathRoot:
-        _widget = const SplashScreen();
+        widget = const SplashScreen();
         break;
       case AuthRoutes.pathRoot:
-        _widget = const AuthRoutes();
+        widget = const AuthRoutes();
         break;
       default:
-        _widget = const AuthRoutes();
+        widget = const AuthRoutes();
         break;
     }
 
-    return MaterialPageRoute(builder: (BuildContext context) => _widget);
+    return MaterialPageRoute(builder: (BuildContext context) => widget);
   }
 }
